Create blackstar guitar jsx
This was a guitar made in illustrator specifically to editable in Framer. Putting code aside, Framer requires that all of your actual paths be on the position in the tree... like a group with a group.  If there are paths in different hierarchies, it gets screwy. But this is one of those SVGs (that's editable in Framer) run through svg2jsx.com